Only unlink the output file on error if it exists
If the cleanup option is being used, then the failed output file may
already have been removed.

+++ python/pod5/src/pod5/tools/pod5_recover.py
@@ -3,6 +3,7 @@ Tool for recovering truncated pod5 files
 """
 import dataclasses
 import typing
+from os.path import exists
 from pathlib import Path
 
 import lib_pod5 as p5b
@@ -79,7 +80,8 @@ def recover_pod5(
         except RuntimeError as error:
             recovered_data.files_with_errors += 1
             print(f"{dest} - Recovery failed - {str(error)}")
-            dest.unlink()
+            if exists(dest):
+                dest.unlink()
 
     print(
         f"Recovered {recovered_data.signal_rows} signal rows, "
